threads.h: modernize ThreadRunner (size_t, uint64_t, enum struct, ref)
Four interrelated changes to ThreadRunner:

- worker() now takes thread_s by reference instead of pointer.
  std::thread(worker, std::ref(tip)) is required because std::thread
  decays its arguments by default; thread_s holds non-copyable
  std::mutex / std::condition_variable / std::thread members, so
  passing tip directly would not compile. Added a comment explaining
  why std::ref is needed.
- thread_s::work was an int64_t with magic values (1: work, 0: wait,
  -1: quit) documented only in a trailing comment. Replaced with
  enum struct Work_state { wait, work, quit } so the state is
  type-checked and self-documenting.
- ThreadRunner thread_count: int → std::size_t. Drops the
  static_cast<int>(parameters.opt_threads) at every call site, plus
  the now-redundant assert(opt_threads <= int_max) checks (size_t is
  wider than the int64_t opt_threads field, and opt_threads is
  validated >= 1 in cli.cc).
- thread_id: int64_t → uint64_t. Propagated through the
  std::function<void(...)> signature, the lambdas in algo.cc /
  algod1.cc / qgram.cc / scan.cc, and the search_worker_core() /
  qgram_worker() / network_thread() / mark_light_thread() /
  check_heavy_thread() declarations. std::next() calls that consume
  thread_id now wrap it in static_cast<std::ptrdiff_t> per
  CLAUDE.md guidance on narrowing/widening conversions.

+++ src/scan.cc
@@ -161,8 +161,8 @@ auto search_getwork(struct Search_state & state,
 
 auto search_worker_core(struct Parameters const & parameters,
                         Data const & data,
-                        const int64_t thread_id, struct Search_state & state) -> void {
-  auto & thread_data = *std::next(state.search_data, thread_id);
+                        const uint64_t thread_id, struct Search_state & state) -> void {
+  auto & thread_data = *std::next(state.search_data, static_cast<std::ptrdiff_t>(thread_id));
   search_init(thread_data, state.query);
   while (search_getwork(state, thread_data.target_count, thread_data.target_index)) {
     search_chunk(parameters, data, thread_data, state, state.master_bits);
